more Psf testing

+++ diffimTests/tasks.py
@@ -124,7 +124,6 @@ def doForcedPhotometry(centroids, exposure, transientsOnly=False, asDF=False):
 # thresholdPolarity: 'both', 'positive', 'negative'
 def doDetection(exp, threshold=5.0, thresholdType='pixel_stdev', thresholdPolarity='positive', doSmooth=True,
                 doMeasure=True, asDF=False):
-    #print 'HERE: NEW DETECTION'
     # Modeled from meas_algorithms/tests/testMeasure.py
     schema = afwTable.SourceTable.makeMinimalSchema()
     config = measAlg.SourceDetectionTask.ConfigClass()
@@ -177,11 +176,10 @@ def doDetection(exp, threshold=5.0, thresholdType='pixel_stdev', thresholdPolari
     return sources
 
 
-def doMeasurePsf(exp, measurePsfAlg='psfex', detectThresh=10.0, startSize=0.01, spatialOrder=1):
+def doMeasurePsf(exp, measurePsfAlg='psfex', detectThresh=10.0, startSize=0.01, spatialOrder=1,
+                 psfMeasureConfig=None):
     # The old (meas_algorithms) SdssCentroid assumed this by default if it
     # wasn't specified; meas_base requires us to be explicit.
-    #print 'HERE: NEW PSF MEASURE'
-
     if exp.getPsf() is not None:  # if possible, use given PSF FWHM/2 to start.
         shape = exp.getPsf().computeImage().getDimensions()
         startSize = exp.getPsf().computeShape().getDeterminantRadius() / 2.
@@ -195,34 +193,38 @@ def doMeasurePsf(exp, measurePsfAlg='psfex', detectThresh=10.0, startSize=0.01, 
     im -= np.median(im.getArray())  # why did I do this?  seems to help sometimes.
 
     # Using 'stdev' seems to work better than 'pixel_stdev' which is the default:
-    sources = doDetection(exp, threshold=detectThresh, thresholdType='stdev', thresholdPolarity='positive')
+    sources = doDetection(exp, threshold=detectThresh, thresholdType='stdev', thresholdPolarity='positive',
+                          doMeasure=True)
+
     #print 'N SOURCES:', len(sources)
-    config = measurePsf.MeasurePsfConfig()
     schema = afwTable.SourceTable.makeMinimalSchema()
+    config = psfMeasureConfig
+    if config is None:  # allow user to provide config
+        config = measurePsf.MeasurePsfConfig()
 
-    if measurePsfAlg is 'psfex':
-        try:
-            import lsst.meas.extensions.psfex.psfexPsfDeterminer
-            config.psfDeterminer['psfex'].spatialOrder = spatialOrder  # 2 is default, 0 seems to kill it
-            config.psfDeterminer['psfex'].recentroid = True
-            config.psfDeterminer['psfex'].sizeCellX = 256  # default is 256
-            config.psfDeterminer['psfex'].sizeCellY = 256
-            config.psfDeterminer['psfex'].samplingSize = 1  # default is 1
-            config.psfDeterminer.name = 'psfex'
-        except ImportError as e:
-            print "WARNING: Unable to use psfex: %s" % e
-            measurePsfAlg = 'pca'
+        if measurePsfAlg is 'psfex':
+            try:
+                import lsst.meas.extensions.psfex.psfexPsfDeterminer
+                config.psfDeterminer['psfex'].spatialOrder = spatialOrder  # 2 is default, 0 seems to kill it
+                config.psfDeterminer['psfex'].recentroid = False #True
+                config.psfDeterminer['psfex'].sizeCellX = 128  # default is 256
+                config.psfDeterminer['psfex'].sizeCellY = 128
+                config.psfDeterminer['psfex'].samplingSize = 1  # default is 1
+                config.psfDeterminer.name = 'psfex'
+            except ImportError as e:
+                print "WARNING: Unable to use psfex: %s" % e
+                measurePsfAlg = 'pca'
 
-    if measurePsfAlg is 'pca':
-        config.psfDeterminer['pca'].sizeCellX = 128
-        config.psfDeterminer['pca'].sizeCellY = 128
-        config.psfDeterminer['pca'].spatialOrder = spatialOrder
-        config.psfDeterminer['pca'].nEigenComponents = 3
-        #config.psfDeterminer['pca'].tolerance = 1e-1
-        #config.starSelector['objectSize'].fluxMin = 500.
-        #config.psfDeterminer['pca'].constantWeight = False
-        #config.psfDeterminer['pca'].doMaskBlends = False
-        config.psfDeterminer.name = "pca"
+        if measurePsfAlg is 'pca':
+            config.psfDeterminer['pca'].sizeCellX = 128
+            config.psfDeterminer['pca'].sizeCellY = 128
+            config.psfDeterminer['pca'].spatialOrder = spatialOrder
+            config.psfDeterminer['pca'].nEigenComponents = 3
+            #config.psfDeterminer['pca'].tolerance = 1e-1
+            #config.starSelector['objectSize'].fluxMin = 500.
+            #config.psfDeterminer['pca'].constantWeight = False
+            #config.psfDeterminer['pca'].doMaskBlends = False
+            config.psfDeterminer.name = "pca"
 
     psfDeterminer = config.psfDeterminer.apply()
     task = measurePsf.MeasurePsfTask(schema=schema, config=config)
